refactor: relocated ValidateEmail into util package
