+++ internal/service/util.go
@@ -1,22 +1,8 @@
 package service
 
 import (
-	"regexp"
-
 	"golang.org/x/crypto/bcrypt"
 )
-
-func validateEmail(email string) error {
-	if email == "" {
-		return ErrEmptyValue
-	}
-	pattern := `^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$`
-	regex := regexp.MustCompile(pattern)
-	if !regex.MatchString(email) {
-		return ErrInvalidEmail
-	}
-	return nil
-}
 
 // hashPasswd generates a hashed version of the password using bcrypt
 func hashPasswd(passwd string) (string, error) {
