+++ internal/util/error.go
@@ -1,0 +1,8 @@
+package util
+
+import "errors"
+
+var (
+	ErrEmptyValue   = errors.New("empty value")
+	ErrInvalidEmail = errors.New("invalid email")
+)
